payment fixure reorganization
