+++ src/payments/enums/billing-interval.enum.ts
@@ -1,0 +1,5 @@
+export enum BillingInterval {
+  MONTHLY = 'MONTHLY',
+  QUARTERLY = 'QUARTERLY',
+  YEARLY = 'YEARLY',
+}
