+++ src/payments/enums/payment-method.enum.ts
@@ -1,0 +1,5 @@
+export enum PaymentMethod {
+  STRIPE = 'STRIPE',
+  PAYPAL = 'PAYPAL',
+  CRYPTO = 'CRYPTO',
+}
