+++ src/payments/enums/subscription-status.enum.ts
@@ -1,0 +1,7 @@
+export enum SubscriptionStatus {
+  ACTIVE = 'ACTIVE',
+  CANCELLED = 'CANCELLED',
+  PAST_DUE = 'PAST_DUE',
+  UNPAID = 'UNPAID',
+  TRIAL = 'TRIAL',
+}
