+++ src/payments/subscriptions/subscriptions.service.spec.ts
@@ -1,7 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { SubscriptionsService } from './subscriptions.service';
 import { getRepositoryToken } from '@nestjs/typeorm';
-import { Subscription, SubscriptionStatus, BillingInterval } from '../entities/subscription.entity';
+import { Subscription } from '../entities/subscription.entity';
+import { SubscriptionStatus, BillingInterval } from '../enums';
 import { StripeService } from '../providers/stripe.service';
 import { Repository } from 'typeorm';
